fix: coerce pgvector float32 to python float at the persistence boundary
On PostgreSQL the pgvector adapter returns numpy float32 arrays. Vector.process_
result_value passed them through unchanged, so list(model.vector) produced lists
of np.float32 and cosine_similarity's sum(x*y ...) yielded np.float32 scores that
pydantic v2 cannot JSON-serialize -> "Object of type float32 is not JSON
serializable" 500s on retrieval/context/agent endpoints. SQLite stores JSON
python floats, so the offline suite never reproduced it.

Fix at the source: Vector.process_result_value now coerces every read vector to
plain python floats ([float(x) for x in value]) on all dialects, eliminating
numpy downstream. Defense-in-depth: cosine_similarity wraps its result in float().

Reproduced pre-fix (score type float32, json.dumps raises the exact error) and
verified post-fix. 27 passed across retrieval/embedding suites + new
test_vector_float32_serialization (4 tests simulating the pgvector numpy path).

+++ backend/app/application/services/retrieval/scoring.py
@@ -29,7 +29,9 @@ def cosine_similarity(a: list[float], b: list[float]) -> float:
     norm_b = math.sqrt(sum(y * y for y in b))
     if norm_a == 0.0 or norm_b == 0.0:
         return 0.0
-    return dot / (norm_a * norm_b)
+    # float() guard: if an embedding ever arrives as numpy (e.g. pgvector
+    # float32), keep the score a plain Python float so it stays JSON-serializable.
+    return float(dot / (norm_a * norm_b))
 
 
 def recency_score(updated_at: datetime, now: datetime, half_life_days: float) -> float:
